refactor: reorganized src folder to reduce clutter and improve structure
It was starting to get messy, so I split things into features, api, and layout.
Hopefully this structure will make the project easier to work with going forward
