+++ src/features/admin/AdminTables.tsx
@@ -1,10 +1,10 @@
-import { useEffect, useState, Table, Button, Modal, Form } from "../index";
+import { useEffect, useState, Table, Button, Modal, Form } from "../../index";
 import {
   getTables,
   updateTables,
   deleteTables,
   createTables,
-} from "../utils/tables";
+} from "../../api/tables";
 
 type Tables = {
   id: number;
