+++ src/features/auth/useAuth.tsx
@@ -1,4 +1,4 @@
-import { useContext } from "./index";
+import { useContext } from "../../index";
 import { AuthContext } from "./AuthContext";
 
 export function useAuth() {
